pyproject.toml
[project]
name = "bluexkye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.116.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-faiss>=0.1.1",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
